refactor(github): refactor migrationscripts
closes #3558
